requirements_conda.txt
absl-py==2.1.0
accelerate==1.3.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cpm-kernels==1.0.11
cycler==0.12.1
datasets==3.2.0
dill==0.3.8
distro==1.9.0
einops==0.8.0
evaluate==0.4.3
exceptiongroup==1.2.2
filelock==3.17.0
fire==0.7.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.9.0
func_timeout==4.3.5
fuzzywuzzy==0.18.0
gradio_client==1.7.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.24.7
-e git+ssh://git@github.com/open-compass/human-eval.git@8d2b96aee463f2aed3c68d5d9ca2dcd24dc1e1aa#egg=human_eval
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.6.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
kiwisolver==1.4.8
langdetect==1.0.9
Levenshtein==0.26.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mmengine-lite==0.10.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.0
OpenCC==1.1.9
-e git+https://github.com/open-compass/opencompass.git@862bf78464ece6b711ac3b4f6ee0337f2471034a#egg=opencompass
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
prettytable==3.14.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pydantic==2.10.6
pydantic_core==2.27.2
pyext==0.7
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.5.0
scipy==1.15.1
seaborn==0.13.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timeout-decorator==0.5.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.2
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
triton==3.2.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
